WIP: Created function WriteGraphSeriesToStatsCollector
Similar to the existing function WriteGraphSeries which writes the
graphs' data to the DB, this function is aimed at writing the data to
the defined Stats Collector instead.

This way, the data does not need to be saved and read from the database
but is rather handled by the stats collector.